Move ProposedService and builder into own file
Moves ProposedService into
store::proposed_service::{ProposedService, ProposedServiceBuilder}

Structs are still exposed only in store:: via pub use. This change
also requires getters to be implemented on the structs to be used
throughout the store implementations.
